Move over some low level collision check from fem/dev.
