+++ geometry/proximity/BUILD.bazel
@@ -21,6 +21,8 @@ drake_cc_package_library(
         ":collisions_exist_callback",
         ":contact_surface_utility",
         ":deformable_contact_internal",
+        ":deformable_contact_surface",
+        ":deformable_rigid_contact_pair",
         ":deformable_volume_mesh",
         ":distance_to_point_callback",
         ":distance_to_shape_callback",
@@ -153,14 +155,13 @@ drake_cc_library(
 )
 
 drake_cc_library(
-    name = "deformable_contact_internal",
-    srcs = ["deformable_contact_internal.cc"],
-    hdrs = ["deformable_contact_internal.h"],
+    name = "deformable_contact_geometries",
+    srcs = ["deformable_contact_geometries.cc"],
+    hdrs = ["deformable_contact_geometries.h"],
     deps = [
         ":deformable_volume_mesh",
         ":hydroelastic_internal",
         ":make_sphere_mesh",
-        ":obj_to_surface_mesh",
         ":proximity_utilities",
         ":tessellation_strategy",
         ":triangle_surface_mesh",
@@ -170,6 +171,46 @@ drake_cc_library(
         "//geometry:proximity_properties",
         "//geometry:shape_specification",
         "//multibody/fem:mesh_utilities",
+    ],
+)
+
+drake_cc_library(
+    name = "deformable_contact_internal",
+    srcs = ["deformable_contact_internal.cc"],
+    hdrs = ["deformable_contact_internal.h"],
+    deps = [
+        "//geometry/query_results:deformable_contact_data",
+        ":deformable_contact_geometries",
+        ":deformable_rigid_contact_pair",
+        ":proximity_utilities",
+        "//common:essential",
+        "//geometry:geometry_ids",
+        "//geometry:proximity_properties",
+        "//geometry:shape_specification",
+    ],
+)
+
+
+drake_cc_library(
+    name = "deformable_contact_surface",
+    srcs = ["deformable_contact_surface.cc"],
+    hdrs = ["deformable_contact_surface.h"],
+    deps = [
+        ":bvh",
+        ":deformable_volume_mesh",
+        ":posed_half_space",
+        ":triangle_surface_mesh",
+        "//common:default_scalars",
+    ],
+)
+
+drake_cc_library(
+    name = "deformable_rigid_contact_pair",
+    hdrs = ["deformable_rigid_contact_pair.h"],
+    deps = [
+        ":deformable_contact_surface",
+        "//geometry:geometry_ids",
+        "//math:geometric_transform",
     ],
 )
 
@@ -825,6 +866,17 @@ drake_cc_googletest(
         "//common/test_utilities:expect_no_throw",
         "//common/test_utilities:expect_throws_message",
         "@fmt",
+    ],
+)
+
+drake_cc_googletest(
+    name = "deformable_contact_surface_test",
+    deps = [
+        ":deformable_contact_surface",
+        "//common/test_utilities:eigen_matrix_compare",
+        "//geometry/proximity:triangle_surface_mesh",
+        "//geometry/proximity:volume_mesh",
+        "//math:autodiff",
     ],
 )
